Drop get_node function

+++ saleor/graphql/core/mutations.py
@@ -11,7 +11,7 @@ from graphql_jwt.exceptions import GraphQLJWTError, PermissionDenied
 from ...account import models
 from ..account.types import User
 from ..file_upload.types import Upload
-from ..utils import get_node, get_nodes
+from ..utils import get_nodes
 from .types.common import Error
 from .utils import snake_to_camel_case
 
@@ -52,28 +52,33 @@ class BaseMutation(graphene.Mutation):
 
     @classmethod
     def add_error(cls, errors, field, message):
-        """Add an error to the errors list.
+        """Add a mutation user error.
 
-        `errors` is the list of errors that happened during execution of the
-        mutation. `field` is the name of model field the error is related
-        to. `None` is allowed and it indicates that the error is a general one
-        and not related to any of the model fields. `message` is the actual
-        error message.
+        `errors` is the list of errors that happened during the execution of
+        the mutation. `field` is the name of an input field the error is
+        related to. `None` value is allowed and it indicates that the error
+        is general and is not related to any of the input fields. `message`
+        is the actual error message to be returned in the response.
 
-        As a result of this method, the `errors` argument is updated with an
-        Error object to be returned in mutation result.
+        As a result of this method, the `errors` list is updated with an Error
+        object to be returned as mutation result.
         """
         field = snake_to_camel_case(field)
         errors.append(Error(field=field, message=message))
 
     @classmethod
-    def get_node_or_error(cls, info, id, errors, field, only_type=None):
-        instance = None
+    def get_node_or_error(cls, info, global_id, errors, field, only_type=None):
+        node = None
         try:
-            instance = get_node(info, id, only_type)
-        except GraphQLError as e:
-            cls.add_error(field=field, message=str(e), errors=errors)
-        return instance
+            node = graphene.Node.get_node_from_global_id(
+                info, global_id, only_type)
+        except (AssertionError, GraphQLError) as e:
+            cls.add_error(errors, field, str(e))
+        else:
+            if node is None:
+                message = "Couldn't resolve to a node: %s" % global_id
+                cls.add_error(errors, field, message)
+        return node
 
     @classmethod
     def get_nodes_or_error(cls, ids, errors, field, only_type=None):
@@ -114,12 +119,6 @@ class ModelMutation(BaseMutation):
             arguments=arguments, fields=fields)
 
     @classmethod
-    def _check_type(cls, field, target_type):
-        if hasattr(field.type, 'of_type'):
-            return field.type.of_type == target_type
-        return field.type == target_type
-
-    @classmethod
     def clean_input(cls, info, instance, input, errors):
         """Clean input data received from mutation arguments.
 
@@ -132,31 +131,44 @@ class ModelMutation(BaseMutation):
         Override this method to provide custom transformations of incoming
         data.
         """
+
+        def is_list_of_ids(field):
+            return (
+                isinstance(field.type, graphene.List) and
+                    field.type.of_type == graphene.ID)
+
+        def is_id_field(field):
+            return (
+                field.type == graphene.ID or
+                    isinstance(field.type, graphene.NonNull) and
+                        field.type.of_type == graphene.ID)
+
+        def is_upload_field(field):
+            if hasattr(field.type, 'of_type'):
+                return field.type.of_type == Upload
+            return field.type == Upload
+
         InputCls = getattr(cls.Arguments, 'input')
         cleaned_input = {}
+
         for field_name, field in InputCls._meta.fields.items():
             if field_name in input:
                 value = input[field_name]
-                # FIXME: maybe we could have custom input field type that takes
-                # the type of IDs
-                # e.g. graphene.IdList(graphene.ID, type=Product).
 
                 # handle list of IDs field
-                if value is not None and (
-                    isinstance(field.type, graphene.List)) and (
-                        field.type.of_type == graphene.ID):
+                if value is not None and is_list_of_ids(field):
                     instances = cls.get_nodes_or_error(
-                        value, errors=errors, field=field.name) if value else []
+                        value, errors=errors, field=field_name) if value else []
                     cleaned_input[field_name] = instances
 
                 # handle ID field
-                elif value is not None and field.type == graphene.ID:
+                elif value is not None and is_id_field(field):
                     instance = cls.get_node_or_error(
-                        info, value, errors=errors, field=field.name)
+                        info, value, errors=errors, field=field_name)
                     cleaned_input[field_name] = instance
 
                 # handle uploaded files
-                elif value is not None and cls._check_type(field, Upload):
+                elif value is not None and is_upload_field(field):
                     value = info.context.FILES.get(value)
                     cleaned_input[field_name] = value
 
@@ -255,7 +267,8 @@ class ModelMutation(BaseMutation):
         # Initialize model instance based on presence of `id` attribute.
         if id:
             model_type = registry.get_type_for_model(cls._meta.model)
-            instance = get_node(info, id, only_type=model_type)
+            instance = cls.get_node_or_error(
+                info, id, errors, 'id', model_type)
         else:
             instance = cls._meta.model()
 
@@ -279,9 +292,15 @@ class ModelDeleteMutation(ModelMutation):
         if not cls.user_is_allowed(info.context.user, data):
             raise PermissionDenied()
 
+        errors = []
         node_id = data.get('id')
         model_type = registry.get_type_for_model(cls._meta.model)
-        instance = get_node(info, node_id, only_type=model_type)
+        instance = cls.get_node_or_error(
+            info, node_id, errors, 'id', model_type)
+
+        if errors:
+            return cls(errors=errors)
+
         db_id = instance.id
         instance.delete()
